requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
cryptography==41.0.4
pytz==2023.3
flask-apscheduler==1.13.1
